siglip2-multilabel/requirements.txt
absl-py==2.3.1
accelerate==1.12.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
airportsdata==20250909
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
apache-tvm-ffi==0.1.6
apex @ file:///tmp/apex
astor==0.8.1
asttokens==3.0.0
attrdict==2.0.1
attrs==25.4.0
audioread==3.1.0
auto_gptq==0.7.1
av==16.0.1
binpacking==1.5.2
bitsandbytes==0.49.0
blake3==1.0.8
blobfile==3.0.0
brotli==1.2.0
build==1.3.0
cachetools==6.2.4
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
compressed-tensors==0.11.0
contourpy==1.3.3
cpm-kernels==1.0.11
crcmod==1.7
cryptography==46.0.3
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cupy-cuda12x==13.6.0
cycler==0.12.1
dacite==1.9.2
datasets==3.6.0
decorator==5.2.1
decord==0.6.0
decord2==3.0.0
deepspeed==0.17.6
depyf==0.19.0
diffusers==0.36.0
dill==0.3.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
dotenv==0.9.9
easydict==1.13
editdistance==0.8.1
einops==0.8.1
email-validator==2.3.0
et_xmlfile==2.0.0
evalscope==1.4.0
evaluate==0.4.6
executing==2.2.1
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
fastrlock==0.8.3
ffmpy==1.0.0
filelock==3.20.0
fire==0.7.1
flash_attn==2.8.3
flashinfer-cubin==0.5.2
flashinfer-python==0.5.2
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.3.0
funasr==1.2.9
func-timeout==4.3.5
future==1.0.0
fuzzywuzzy==0.18.0
gekko==1.3.0
genson==1.3.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
google-genai==1.56.0
gradio==5.32.1
gradio_client==1.10.2
groovy==0.1.2
grpcio==1.75.1
grpcio-health-checking==1.75.1
grpcio-reflection==1.75.1
grpcio-tools==1.75.1
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
human-eval==1.0.3
hydra-core==1.3.2
icecream==2.1.8
identify==2.6.15
idna==3.11
ImageIO==2.37.2
immutabledict==4.2.2
importlib_metadata==8.7.1
interegular==0.3.3
ipdb==0.13.13
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jaconv==0.4.1
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.3
json5==0.12.1
json_repair==0.54.3
jsonlines==4.0.0
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaldiio==2.18.1
kiwisolver==1.4.9
langdetect==1.0.9
lark==1.2.2
latex2sympy2_extended==1.10.2
lazy_loader==0.4
Levenshtein==0.27.3
librosa==0.11.0
liger_kernel==0.6.4
lightning-utilities==0.15.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
lmdeploy==0.10.1
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
math-verify==0.8.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
megatron-core @ git+https://github.com/NVIDIA/Megatron-LM.git@0d7e02bd26d3eb347ad8d1bef625002babfa6df7
mistral_common==1.8.8
ml_dtypes==0.5.4
mmengine-lite==0.10.7
modelscope @ file:///tmp/modelscope
more-itertools==10.8.0
mpmath==1.3.0
ms-agent==1.5.2
ms-opencompass==0.1.6
ms-swift @ file:///tmp/ms-swift
ms-vlmeval==0.0.19
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
nltk==3.9.2
nodeenv==1.10.0
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.17.0
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.2.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.9.79
nvitop==1.6.1
omegaconf==2.3.0
onnx==1.20.0
onnx-ir==0.1.13
onnxscript==0.5.7
openai==2.6.1
openai-harmony==0.0.4
openai-whisper==20250625
OpenCC==1.1.9
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
optimum==2.1.0
orjson==3.11.5
oss2==2.19.1
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
partial-json-parser==0.2.1.1.post7
peft==0.17.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.1
ply==3.11
pooch==1.8.2
portalocker==3.2.0
pre_commit==4.5.1
prettytable==3.17.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pybind11==3.0.1
pycountry==24.6.1
pycparser==2.23
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.10
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydub==0.25.1
pyecharts==2.0.9
Pygments==2.19.2
pylatexenc==2.10
pynndescent==0.5.13
pynvml==13.0.1
pyparsing==3.3.1
pypinyin==0.55.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-Levenshtein==0.27.3
python-multipart==0.0.21
pytorch-wpe==0.0.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qwen-omni-utils==0.0.8
qwen-vl-utils==0.0.14
rank-bm25==0.2.2
RapidFuzz==3.14.3
ray==2.53.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==13.9.4
rich-toolkit==0.17.1
rignore==0.7.6
rouge==1.0.1
rouge-chinese==1.0.3
rouge-score==0.1.2
rpds-py==0.30.0
rsa==4.9.1
ruff==0.14.10
sacrebleu==2.5.1
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.2.0
sentencepiece==0.2.1
sentry-sdk==2.48.0
setproctitle==1.3.7
sgl-kernel==0.3.17.post1
sglang==0.5.5.post3
shellingham==1.5.4
shortuuid==1.0.13
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
starlette==0.50.0
sty==1.0.6
swanlab==0.7.5
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.2.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tiktoken==0.12.0
timeout-decorator==0.5.0
timm==1.0.16
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.8.0+cu129
torch-complex==0.4.4
torch_memory_saver==0.0.9
torchao==0.9.0
torchaudio==2.8.0+cu129
torchmetrics==1.8.2
torchvision==0.23.0+cu129
tqdm==4.67.1
traitlets==5.14.3
transformer-engine-torch==2.10.0
transformer_engine==2.10.0
transformer_engine_cu12==2.10.0
transformers==4.57.3
transformers-stream-generator==0.0.5
triton==3.4.0
trl==0.24.0
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.9.post2
urllib3==2.6.2
uvicorn==0.40.0
uvloop==0.21.0
validators==0.35.0
virtualenv==20.35.4
vllm==0.11.0
wandb==0.23.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.4
word2number==1.1
wrapt==2.0.1
xformers==0.0.32.post1
xgrammar==0.1.25
xlsxwriter==3.2.9
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zhconv==1.4.3
zipp==3.23.0
zstandard==0.25.0
